+++ CHANGELOG.md
@@ -13,6 +13,31 @@ Modifications by (in alphabetical order):
 * B. Reuter, ECMWF, UK
 * J. Tiira, University of Helsinki, Finland
 * P. Vitt, University of Siegen, Germany
+* A. Voysey, UK Met Office
+
+11/01/2021 PR #285 for #283. Removes Structure_Constructor_2 class to fix
+           problem with Expr matching.
+
+11/01/2021 PR #237 for #236. Fix for slow parsing and printing of code.
+
+23/12/2020 PR #278 for #277. Added github actions CI file, badges and documentation.
+
+18/12/2020 PR #275 for #274. Adds an example script that uses fparser2 to
+           generate Makefile dependency rules for a set of Fortran source
+           files.
+
+30/11/2020 PR #272 for #271. Bug fix to module_in_file() to ensure that the
+           encoding is always set to UTF-8 when reading a file.
+
+25/10/2020 PR #256 for #252. Fixes a bug in the parsing of an array constructor.
+
+03/06/2020 PR #263 for #262. Fixes bug in fparser2 logging a 'critical'
+           error when it reaches the end of a file.
+
+03/06/2020 PR #249. Adds support for Fortran2008 CONTIGUOUS and
+           CODIMENSION keywords.
+
+## Release 0.0.11 (15/05/2020) ##
 
 09/04/2020 PR #254. Fix for >1 character pre-processor macro identifiers.